--- src/components/Message.tsx
@@ -1,217 +1,0 @@
-import React from 'react';
-import { IMessage } from '../types';
-import { AnthropicMessage } from './AnthropicMessage';
-import UserMessage from './UserMessage';
-import SystemMessage from './SystemMessage';
-import { OpenAIMessage } from './OpenAIMessage';
-import OpenAI from 'openai';
-
-interface MessageProps {
-  message: IMessage;
-}
-
-const Message: React.FC<MessageProps> = ({ message }) => {
-  if (message?.service === 'openai') {
-    return <OpenAIMessage key={message?.id} message={(message as OpenAI.ChatCompletion)} />; // TODO: Create <OpenAIMessage message={message} />;
-  } else if (message?.service === 'anthropic') {
-    return <AnthropicMessage key={message?.id} message={message} />;
-  } else if (message?.role === 'system') {
-    return <SystemMessage key={message?.id} message={message} />;
-  } else if (message?.role === 'user') {
-    return <UserMessage key={message?.id} message={message} />;
-  } else {
-    return <div>{JSON.stringify(message, null, 2)}</div>;
-  }
-
-  // const renderContent = (item: ContentItem): JSX.Element => {
-  //   if (isTitle(item)) {
-  //     return <h3 style={{ whiteSpace: 'pre-wrap' }} key={item.id} className="text-lg font-bold">{item.title}</h3>;
-  //   } else if (isText(item)) {
-  //     try {
-  //       if (typeof item.text !== 'string') throw new Error('Value is not a string');
-  //       if (item.text === '') return <p key={item.id}></p>;
-  //       if (item.text.startsWith('[') && item.text.endsWith(']')) {
-  //         const parsedText = JSON.parse(item.text);
-  //         if (Array.isArray(parsedText)) {
-  //           return (
-  //             <div className="space-y-2" key={item.text}>{parsedText.map((item) => renderContent(item))}</div>
-  //           );
-  //         } else {
-  //           return <p style={{ whiteSpace: 'pre-wrap' }} key={item.text}>{item.text}</p>;
-  //         }
-  //       } else if (item.text.startsWith('{') && item.text.endsWith('}')) {
-  //         return renderContent(JSON.parse(item.text));
-  //       } else {
-  //         return <p style={{ whiteSpace: 'pre-wrap' }} key={item.id}>{item.text}</p>;
-  //       }
-  //     } catch (error) {
-  //       return (
-  //         <div key={item.id} className="space-y-2">
-  //           <p style={{ whiteSpace: 'pre-wrap' }}>{JSON.stringify(item.text, null, 2)}</p>
-  //           <p style={{ whiteSpace: 'pre-wrap' }} className="text-sm text-red-500">Error: {(error as Error).message ? (error as Error).message : JSON.stringify(error, null, 2)}</p>
-  //         </div>
-  //       );
-  //     }
-  //   } else if (isUpdatedProjectState(item)) {
-  //     return (
-  //       <MessageProjectStateUpdates key={item.id} projectStateUpdates={item?.project_state_updates} />
-  //     );
-  //   } else if (isCode(item)) {
-  //     const [code, fileExt, filePath, description] = item.code;
-  //     if (fileExt === 'json') {
-  //       try {
-  //         const parsedCode = JSON.parse(code);
-  //         if (Array.isArray(parsedCode)) {
-  //           return (
-  //             <div className="space-y-2" key={item.id}>{parsedCode.map((item) => renderContent(item))}</div>
-  //           );
-  //         } else if (parsedCode?.project_state_updates) {
-  //           return <MessageProjectStateUpdates key={item.id} projectStateUpdates={parsedCode?.project_state_updates} />;
-  //         }
-  //       } catch (error) {
-  //         return (
-  //           <div key={item.id} className="space-y-2">
-  //             <p style={{ whiteSpace: 'pre-wrap' }}>{item.code}</p>
-  //             <p style={{ whiteSpace: 'pre-wrap' }} className="text-sm text-red-500">Error: {(error as Error).message ? (error as Error).message : JSON.stringify(error, null, 2)}</p>
-  //           </div>
-  //         );
-  //       }
-  //     }
-  //     return (
-  //       <div key={item.id} className="space-y-2">
-  //         {renderCodeBlock(code, fileExt || undefined)}
-  //         {description && <p style={{ whiteSpace: 'pre-wrap' }} className="text-sm text-gray-500">{description}</p>}
-  //         <div className="flex items-center space-x-2">
-  //           <button
-  //             onClick={async () => { await writeFile(code, filePath); }}
-  //             className="bg-blue-500 hover:bg-blue-700 text-white font-bold px-3 rounded-full"
-  //           >
-  //             Write to file
-  //           </button>
-  //           <span style={{ whiteSpace: 'pre-wrap' }} className="text-sm text-gray-500">{filePath || '/'}</span>
-  //         </div>
-  //       </div>
-  //     );
-  //   } else if (isLink(item)) {
-  //     const [url, description] = item.link;
-  //     return (
-  //       <a href={url} style={{ whiteSpace: 'pre-wrap' }} target="_blank" rel="noopener noreferrer" className="text-blue-500 hover:underline">
-  //         {description || url}
-  //       </a>
-  //     );
-  //   };
-  //   return <p key={item?.id} style={{ whiteSpace: 'pre-wrap' }}>{JSON.stringify(item, null, 2)}</p>;
-  // };
-
-  // const parseMessageContent = (content: MessageContent | string): MessageContent | JSX.Element[] => {
-  //   try {
-  //     let parsedContent = content;
-  //     if (typeof content === 'string') {
-  //       parsedContent = JSON.parse(content);
-  //     };
-  //     if (Array.isArray(parsedContent)) {
-  //       return parsedContent as MessageContent;
-  //     }
-  //     return parseTextWithCodeBlocks(JSON.stringify(parsedContent, null, 2)); // TODO: try without JSON.stringify
-  //   } catch (error) {
-  //     return parseTextWithCodeBlocks(content as string); // TODO: try with JSON.stringify
-  //   }
-  // };
-
-  // const parseTextWithCodeBlocks = (text: string): JSX.Element[] => {
-  //   try {
-
-  //     if (typeof text !== 'string') throw new Error('Value is not a string');
-  //     const parts = text.split(/(```[\s\S]*?```)/);
-  //     return parts.map((part, index) => {
-  //       if (part.startsWith('```') && part.endsWith('```')) {
-  //         const [, language, code] = part.match(/```(\w*)\n([\s\S]*?)```/) || [, '', part.slice(3, -3)];
-  //         return (
-  //           <div key={index} className="space-y-2">
-  //             {renderCodeBlock(code.trim(), language)}
-  //             <div className="flex items-center space-x-2">
-  //               <button
-  //                 onClick={async () => writeFile(code.trim(), '')}
-  //                 className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-1 px-2 rounded-md text-sm"
-  //               >
-  //                 Write to file
-  //               </button>
-  //               <span className="text-sm text-gray-500">No path suggested</span>
-  //             </div>
-  //           </div>
-  //         );
-  //       } else {
-  //         return <p key={index} style={{ whiteSpace: 'pre-wrap' }}>{part}</p>;
-  //       }
-  //     });
-  //   } catch (error) {
-  //     const stringifiedText = JSON.stringify(text, null, 2);
-  //     return [(
-  //       <div key={stringifiedText} className="space-y-2">
-  //         <p style={{ whiteSpace: 'pre-wrap' }}>{stringifiedText}</p>
-  //         <p style={{ whiteSpace: 'pre-wrap' }} className="text-sm text-red-500">Error: {(error as Error).message ? (error as Error).message : JSON.stringify(error, null, 2)}</p>
-  //       </div>
-  //     )];
-  //   }
-  // };
-
-  // const renderCodeBlock = (code: string, language: string = 'txt') => {
-  //   return (
-  //     <Editor
-  //       value={code}
-  //       language={language}
-  //       minHeight={24}
-  //       style={{
-  //         marginLeft: 25,
-  //         lineHeight: 1.6,
-  //       }}
-  //       disabled={true}
-  //     />
-  //   );
-  // };
-
-  // const parsedContent = parseMessageContent(message?.content);
-
-  // return (
-  //   <div key={message.id} className={`flex flex-col items-center max-w-max select-text justify-end`}>
-  //     <div className={`flex flex-col p-2 space-y-2 rounded-md bg-blue-600 bg-opacity-20 hover:shadow-md`}>
-  //       {Array.isArray(parsedContent) ? (
-  //         parsedContent.map((item, index) => {
-  //           if (React.isValidElement(item)) {
-  //             return (<div key={`${Date.now()}${index}`} className="flex flex-col mb-4 rounded-md max-w-[800px] space-y-2">{item}</div>);
-  //           }
-  //           return (
-  //             <div key={`${Date.now()}${index}`} className="flex flex-col mb-4 rounded-md max-w-[800px] space-y-2">
-  //               {renderContent(item as ContentItem)}
-  //             </div>
-  //           );
-  //         })
-  //       ) : (
-  //         <p key={parsedContent} style={{ whiteSpace: 'pre-wrap' }} className="mb-4">
-  //           {parsedContent}
-  //         </p>
-  //       )}
-  //       <div className="text-xs opacity-50">
-  //         {new Date((message as IBaseMessage)?.createdAt ?? '').toLocaleString()}
-  //         {(message?.createdAt !== message?.updatedAt) && " (edited)"}
-  //       </div>
-  //       <Accordion
-  //         title="Raw message"
-  //         className=""
-  //         titleClassName="text-xs opacity-50"
-  //         buttonClassName=""
-  //         content={
-  //           <div
-  //             style={{ whiteSpace: 'pre-wrap' }}
-  //             className="text-xs opacity-50"
-  //           >
-  //             {JSON.stringify(message, null, 2)}
-  //           </div>
-  //         }
-  //       />
-  //     </div>
-  //   </div>
-  // );
-};
-
-export default Message;
