+++ src/components/Requirements.tsx
@@ -59,7 +59,6 @@ const Requirements: React.FC = () => {
             placeholder="Enter a requirement..."
             rows={1}
           />
-          {/* {requirement?.update && <span className={`px-2 py-1 text-sm`}>{requirement.update}</span>} */}
         </div>
       ))}
       <Textarea
